Refactor drag_and_drop_files_widget.py to pages/core.
