+++ pages/core/drag_and_drop_files_widget.py
